wow really easy to extend cloud zarr sources
